updated docs. udpates getstudents
