+++ docs/cqrs.md
@@ -1,6 +1,4 @@
-Command Query Responsibility Segregation (CQRS)
------------------------------------------------
-The [CQRS](https://martinfowler.com/bliki/CQRS.html) implementation in this framework was majorly inspired by [these series of posts](https://jimmybogard.com/life-beyond-transactions-implementation-primer/). To understand more about the CQRS pattern and when to use it, please go over [this post](https://martinfowler.com/bliki/CQRS.html).
+The Command Query Responsibility Segregation (CQRS) implementation in this framework was majorly inspired by [these series of posts](https://jimmybogard.com/life-beyond-transactions-implementation-primer/). To understand more about the CQRS pattern and when to use it, please go over [this post](https://martinfowler.com/bliki/CQRS.html).
 
 In order to implement CQRS, you need to do the following:
 
@@ -22,3 +20,8 @@ Like this:
 
 * Configure EF DbContext and MassTransit Bus.
 * Create your domain entity class and inherit [`BaseCqrsEntity`](https://github.com/gokulm/BoltOn/blob/master/src/BoltOn/Cqrs/BaseCqrsEntity.cs), and which will force your entity's Id property to be of type Guid. The `BaseCqrsEntity` class has two properties `EventsToBeProcessed` and `ProcessedEvents`, and two methods `RaiseEvent<TEvent>(TEvent @event` and `ProcessEvent<TEvent>(TEvent @event, Action<TEvent> action)` that facilitates to implement the pattern. 
+* Create your request and handlers, and then use the `Mediator` to process your request. Please refer to [Mediator](../mediator) documentation to create handlers.
+
+The best way to understand the implementation is by looking into [BoltOn.Samples.WebApi](https://github.com/gokulm/BoltOn/tree/master/samples/BoltOn.Samples.WebApi) project's StudentsController and by going thru GET, POST and PUT student endpoints, corresponding requests and their handlers. `Student` and `StudentFlattened` entities inherit `BaseCqrsEntity`. Events get triggered on create and update of Student entity, and they get handled by handlers registered in [BoltOn.Samples.Console](https://github.com/gokulm/BoltOn/tree/master/samples/BoltOn.Samples.Console) project.
+
+Student gets instantiated in `CreateStudentHandler` using the internal constructor (private ctor is to support EF and the internal ctor is to support unit testing but still encapsulate the properties). StudentCreatedEvent event gets triggered in the ctor.
