+++ samples/BoltOn.Samples.Application/Handlers/GetAllStudentsHandler.cs
@@ -2,9 +2,10 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using BoltOn.Data;
 using BoltOn.Mediator.Pipeline;
-using BoltOn.Samples.Application.Abstractions.Data;
 using BoltOn.Samples.Application.DTOs;
+using BoltOn.Samples.Application.Entities;
 
 namespace BoltOn.Samples.Application.Handlers
 {
@@ -14,21 +15,24 @@ namespace BoltOn.Samples.Application.Handlers
 
     public class GetAllStudentsHandler : IRequestAsyncHandler<GetAllStudentsRequest, IEnumerable<StudentDto>>
     {
-        private readonly IStudentRepository _studentRepository;
+        private readonly IRepository<StudentFlattened> _studentRepository;
 
-        public GetAllStudentsHandler(IStudentRepository studentRepository)
+        public GetAllStudentsHandler(IRepository<StudentFlattened> studentRepository)
         {
             _studentRepository = studentRepository;
         }
         
-		public async Task<IEnumerable<StudentDto>> HandleAsync(GetAllStudentsRequest request, CancellationToken cancellationToken = default(CancellationToken))
+		public async Task<IEnumerable<StudentDto>> HandleAsync(GetAllStudentsRequest request, 
+			CancellationToken cancellationToken = default(CancellationToken))
 		{
 			var students = (await _studentRepository.GetAllAsync()).ToList();
 			var studentDtos = from s in students
 							   select new StudentDto
 							   {
+								   Id = s.Id,
 								   FirstName = s.FirstName,
-								   LastName = s.LastName
+								   LastName = s.LastName,
+								   StudentType = s.StudentType
 							   };
 			return studentDtos;
 		}
