+++ samples/BoltOn.Samples.WebApi/Startup.cs
@@ -5,17 +5,10 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using BoltOn.Data.EF;
 using BoltOn.Samples.Application.Handlers;
-using Microsoft.EntityFrameworkCore;
-using BoltOn.Samples.Infrastructure.Data;
 using BoltOn.Samples.Infrastructure.Data.Repositories;
 using BoltOn.Data.CosmosDb;
 using BoltOn.Bus.MassTransit;
-using MassTransit;
-using System;
 using BoltOn.Utilities;
-using BoltOn.Data;
-using BoltOn.Samples.Application.Entities;
-using BoltOn.Bootstrapping;
 
 namespace BoltOn.Samples.WebApi
 {
@@ -46,40 +39,6 @@ namespace BoltOn.Samples.WebApi
 			app.UseMvc();
 			app.ApplicationServices.TightenBolts();
 			appLifetime.ApplicationStopping.Register(() => BoltOnAppCleaner.Clean());
-		}
-	}
-
-	public class RegistrationTask : IRegistrationTask
-	{
-		public void Run(RegistrationTaskContext context)
-		{
-			var container = context.Container;
-			container.AddMassTransit(x =>
-			{
-				x.AddBus(provider => MassTransit.Bus.Factory.CreateUsingRabbitMq(cfg =>
-				{
-					var host = cfg.Host(new Uri("rabbitmq://localhost:5672"), hostConfigurator =>
-					{
-						hostConfigurator.Username("guest");
-						hostConfigurator.Password("guest");
-					});
-				}));
-			});
-
-			container.AddDbContext<SchoolDbContext>(options =>
-			{
-				options.UseSqlServer("Data Source=127.0.0.1;initial catalog=Testing;persist security info=True;User ID=sa;Password=Password1;");
-			});
-
-			//services.AddCosmosDbContext<CollegeDbContext>(options =>
-			//{
-			//    options.Uri = "";
-			//    options.AuthorizationKey = "";
-			//    options.DatabaseName = "";
-			//});
-
-			container.AddTransient<IRepository<Student>, CqrsRepository<Student, SchoolDbContext>>();
-			container.AddTransient<IRepository<StudentType>, Repository<StudentType, SchoolDbContext>>();
 		}
 	}
 }
